requirements.txt
flask>=3.0
python-dotenv>=1.0
requests>=2.31
beautifulsoup4>=4.12
apscheduler>=3.10
lxml>=5.0
